refactor(types): model retail word-sized state
Replace decompiler int/void* residue with fixed-width u32/u8/s16 where
retail ASM reads and writes raw words, bytes, and halfwords.

Rename recycled audio sound slots from audioPtr to soundIDCount, matching
the OtherFX/PlaySound3D_Flags word-handle contract and removing casts at
call sites.

+++ platform/native_audio.c
@@ -513,7 +513,7 @@ internal int NativeAudio_Clamp16(int value)
 	return value;
 }
 
-internal int NativeAudio_VolumeScale(s16 volume)
+internal int NativeAudio_VolumeScale(int volume)
 {
 	if (volume < 0)
 	{
